UHF-10909: Added social media service parser trait.
